SERVER-10275 Use size_t in pathsupport function

+++ src/mongo/db/ops/modifier_bit.cpp
@@ -42,7 +42,7 @@ namespace mongo {
         mutablebson::Document& doc;
 
         // Index in _fieldRef for which an Element exist in the document.
-        int32_t idxFound;
+        size_t idxFound;
 
         // Element corresponding to _fieldRef[0.._idxFound].
         mutablebson::Element elemFound;
@@ -168,7 +168,7 @@ namespace mongo {
         // If the field path is not fully present, then this mod cannot be in place, nor is a
         // noOp.
         if (!_preparedState->elemFound.ok() ||
-            _preparedState->idxFound < static_cast<int32_t>(_fieldRef.numParts() - 1)) {
+            _preparedState->idxFound < (_fieldRef.numParts() - 1)) {
             // If no target element exists, the value we will write is the result of applying
             // the operation to a zero-initialized integer element.
             _preparedState->newValue = apply(SafeNum(static_cast<int>(0)));
@@ -211,7 +211,7 @@ namespace mongo {
         // If there's no need to create any further field part, the $bit is simply a value
         // assignment.
         if (_preparedState->elemFound.ok() &&
-            _preparedState->idxFound == static_cast<int32_t>(_fieldRef.numParts() - 1)) {
+            _preparedState->idxFound == (_fieldRef.numParts() - 1)) {
             return _preparedState->elemFound.setValueSafeNum(_preparedState->newValue);
         }
