+++ src/mongo/db/ops/modifier_set.cpp
@@ -40,7 +40,7 @@ namespace mongo {
         mutablebson::Document& doc;
 
         // Index in _fieldRef for which an Element exist in the document.
-        int32_t idxFound;
+        size_t idxFound;
 
         // Element corresponding to _fieldRef[0.._idxFound].
         mutablebson::Element elemFound;
@@ -163,7 +163,7 @@ namespace mongo {
         // If the field path is not fully present, then this mod cannot be in place, nor is a
         // noOp.
         if (!_preparedState->elemFound.ok() ||
-            _preparedState->idxFound < static_cast<int32_t>(_fieldRef.numParts()-1)) {
+            _preparedState->idxFound < (_fieldRef.numParts()-1)) {
             return Status::OK();
         }
 
@@ -178,7 +178,7 @@ namespace mongo {
         // If the value being $set is the same as the one already in the doc, than this is a
         // noOp.
         if (_preparedState->elemFound.ok() &&
-            _preparedState->idxFound == static_cast<int32_t>(_fieldRef.numParts()-1) &&
+            _preparedState->idxFound == (_fieldRef.numParts()-1) &&
             _preparedState->elemFound.compareWithBSONElement(_val, false /*ignore field*/) == 0) {
             execInfo->noOp = _preparedState->noOp = true;
         }
@@ -192,7 +192,7 @@ namespace mongo {
         // If there's no need to create any further field part, the $set is simply a value
         // assignment.
         if (_preparedState->elemFound.ok() &&
-            _preparedState->idxFound == static_cast<int32_t>(_fieldRef.numParts()-1)) {
+            _preparedState->idxFound == (_fieldRef.numParts()-1)) {
             return _preparedState->elemFound.setValueBSONElement(_val);
         }
